update
foi feito o menu e a mecaninca de ganhar
