+++ Assets/Animations/Menu.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cff5b3e8f391b7c40a4db013e2c08bda
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
