+++ Assets/Animations/Menu/Start.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f22c0917186bfdc4dac67a22c649c781
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
