+++ Assets/Animations/Menu/Text (TMP).controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 95fe236ccbd249f418c3d2b262958339
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
