--- Assets/Scenes/FinalFase.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: d7d1a4cac3f3d314693c2b6a97cf2ccd
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
