--- Assets/Scenes/SecondFase.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: f3f1c7d56765040499b202388b26c519
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
