--- Assets/Scenes/Start.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 8192ecb8053f1aa4ba5c2575cd96dc0e
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
