+++ Assets/Scipts/StartGame.cs
@@ -1,0 +1,12 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.SceneManagement;
+
+public class StartGame : MonoBehaviour
+{
+    public void Game()
+    {
+        UnityEngine.SceneManagement.SceneManager.LoadScene("FirstFase");
+    }
+}
