+++ Assets/Scipts/StartGame.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 175040372f41d254fa79a31d461c3eda
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
