+++ Assets/Scipts/VehicleFinal.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b71f58efa071ba047a88bd01339247ce
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
