--- Assets/Scipts/VehicleFinal1.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ab4b783ffca3f6d469f8c1c79c1e690d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
